% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppTskitTestLinkingTo-package.R
\docType{package}
\name{RcppTskitTestLinking-package}
\alias{RcppTskitTestLinking}
\alias{RcppTskitTestLinking-package}
\title{RcppTskitTestLinking: Testing linking to and importing from RcppTskit}
\description{
This is a demo package that shows how to link to and import from RcppTskit
package, which provides R access to tskit C API.
}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} [copyright holder]

}
\keyword{internal}
