% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtsk_table_collection_summary2}
\alias{rtsk_table_collection_summary2}
\title{Summarise \code{tskit} table collection}
\usage{
rtsk_table_collection_summary2(tc)
}
\arguments{
\item{tc}{an external pointer to a \code{tsk_table_collection_t} object.}
}
\value{
A list.
}
\description{
Summarise \code{tskit} table collection
}
\examples{
ts_file <- system.file("examples", "test.trees", package = "RcppTskit")
tc <- RcppTskit:::tc_load(ts_file)
RcppTskit:::rtsk_table_collection_summary(tc$xptr)
rtsk_table_collection_summary2(tc$xptr)
}
