% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtsk_treeseq_get_num_individuals2}
\alias{rtsk_treeseq_get_num_individuals2}
\title{Get number of individuals in tree sequence}
\usage{
rtsk_treeseq_get_num_individuals2(ts)
}
\arguments{
\item{ts}{an external pointer to a \code{tsk_treeseq_t} object.}
}
\value{
integer number of individuals.
}
\description{
Get number of individuals in tree sequence
}
\examples{
ts_file <- system.file("examples", "test.trees", package = "RcppTskit")
ts <- RcppTskit::ts_load(ts_file)
ts$num_individuals()
RcppTskit:::rtsk_treeseq_get_num_individuals(ts$xptr)
rtsk_treeseq_get_num_individuals2(ts$xptr)
}
