% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_num_individuals_ptr2}
\alias{ts_num_individuals_ptr2}
\title{Get number of individuals in tree sequence}
\usage{
ts_num_individuals_ptr2(ts)
}
\arguments{
\item{ts}{an external pointer to a \code{tsk_treeseq_t} object.}
}
\value{
integer number of individuals.
}
\description{
Get number of individuals in tree sequence
}
\examples{
ts_file <- system.file("examples", "test.trees", package = "RcppTskit")
ts <- RcppTskit::ts_load(ts_file)
ts$num_individuals()
ts_num_individuals_ptr2(ts$pointer)
}
